+++ BitmovinAnalyticsCollector/Classes/AVPlayerV2/AVPlayerAdapter.swift
@@ -29,15 +29,19 @@ class AVPlayerAdapter: CorePlayerAdapter, PlayerAdapter {
     private let errorHandler: ErrorHandler
     private let bitrateDetectionService: BitrateDetectionService
     private let playbackTypeDetectionService: PlaybackTypeDetectionService
+    private let downloadSpeedDetectionService: DownloadSpeedDetectionService
+    private let downloadSpeedMeter: DownloadSpeedMeter
     private let manipulator: AVPlayerEventDataManipulator
     
     init(player: AVPlayer, config: BitmovinAnalyticsConfig, stateMachine: StateMachine) {
         self.player = player
         self.config = config
         self.errorHandler = ErrorHandler()
-        self.bitrateDetectionService = BitrateDetectionService(bitrateLogProvider: AVPlayerBitrateLogProvider(player: player))
+        self.bitrateDetectionService = BitrateDetectionService()
+        self.downloadSpeedMeter = DownloadSpeedMeter()
+        self.downloadSpeedDetectionService = DownloadSpeedDetectionService(downloadSpeedMeter: self.downloadSpeedMeter)
         self.playbackTypeDetectionService = PlaybackTypeDetectionService(player: player)
-        self.manipulator = AVPlayerEventDataManipulator(player: player, playbackTypeDetectionService: playbackTypeDetectionService)
+        self.manipulator = AVPlayerEventDataManipulator(player: player, playbackTypeDetectionService: playbackTypeDetectionService, downloadSpeedMeter: self.downloadSpeedMeter)
         super.init(stateMachine: stateMachine)
     }
     
@@ -84,11 +88,6 @@ class AVPlayerAdapter: CorePlayerAdapter, PlayerAdapter {
         playerKVOList.append(player.observe(\.currentItem, options: [.new, .old, .initial]) {[weak self] (player, change) in
             self?.onPlayerCurrentItemChange(old: change.oldValue ?? nil, new: change.newValue ?? nil)
         })
-        
-        bitrateDetectionService.startMonitoring()
-        bitrateDetectionServiceKVO = bitrateDetectionService.observe(\.videoBitrate, options: [.new, .old]) { [weak self] _, change in
-            self?.onVideoQualityChange(newVideoBitrate: change.newValue ?? nil)
-        }
     }
 
     override public func stopMonitoring() {
@@ -111,10 +110,6 @@ class AVPlayerAdapter: CorePlayerAdapter, PlayerAdapter {
         }
         playerKVOList.removeAll()
         
-        bitrateDetectionService.stopMonitoring()
-        bitrateDetectionServiceKVO?.invalidate()
-        bitrateDetectionServiceKVO = nil
-        
         resetSourceState()
     }
 
@@ -127,7 +122,16 @@ class AVPlayerAdapter: CorePlayerAdapter, PlayerAdapter {
         NotificationCenter.default.addObserver(self, selector: #selector(observeFailedToPlayToEndTime(notification:)), name: NSNotification.Name.AVPlayerItemFailedToPlayToEndTime, object: playerItem)
         NotificationCenter.default.addObserver(self, selector: #selector(observeDidPlayToEndTime(notification:)), name: NSNotification.Name.AVPlayerItemDidPlayToEndTime, object: playerItem)
         manipulator.updateDrmPerformanceInfo(playerItem)
+        
         playbackTypeDetectionService.startMonitoring(playerItem: playerItem)
+        
+        let accessLogProvider = AVPlayerAccessLogProvider(playerItem: playerItem)
+        bitrateDetectionService.startMonitoring(accessLogProvider: accessLogProvider)
+        bitrateDetectionServiceKVO = bitrateDetectionService.observe(\.videoBitrate, options: [.new, .old]) { [weak self] _, change in
+            self?.onVideoQualityChange(newVideoBitrate: change.newValue ?? nil)
+        }
+        
+        downloadSpeedDetectionService.startMonitoring(accessLogProvider: accessLogProvider)
     }
 
     private func stopMonitoringPlayerItem(playerItem: AVPlayerItem) {
@@ -136,7 +140,14 @@ class AVPlayerAdapter: CorePlayerAdapter, PlayerAdapter {
         NotificationCenter.default.removeObserver(self, name: NSNotification.Name.AVPlayerItemFailedToPlayToEndTime, object: playerItem)
         NotificationCenter.default.removeObserver(self, name: NSNotification.Name.AVPlayerItemDidPlayToEndTime, object: playerItem)
         playerItemStatusObserver?.invalidate()
+        
         playbackTypeDetectionService.stopMonitoring(playerItem: playerItem)
+        
+        bitrateDetectionService.stopMonitoring()
+        bitrateDetectionServiceKVO?.invalidate()
+        bitrateDetectionServiceKVO = nil
+        
+        downloadSpeedDetectionService.stopMonitoring()
     }
 
     // AVPlayerItem KVOs