+++ BitmovinAnalyticsCollector/Classes/AVPlayerV2/data/AccessLogProvider.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+protocol AccessLogProvider {
+    func getEvents() -> [AccessLogDto]?
+}