+++ BitmovinAnalyticsCollector/Classes/AVPlayerV2/data/BitrateLogDto.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-class BitrateLogDto {
-    var durationWatched: Double = 0.0
-    var indicatedBitrate: Double = 0.0
-}