+++ BitmovinAnalyticsCollector/Tests/AVPlayerV2Tests/services/BitrateDetectionServiceTest.swift
@@ -12,12 +12,24 @@ import XCTest
 
 class BitrateDetectionServiceTest: XCTestCase {
     
-    func test_detectBitrateChange_should_doNothing_when_bitrateLogIsNil() {
+    func test_detectBitrateChange_should_doNothing_when_accessLogIsNil() {
         // arrange
-        let mockBitrateLogProvider = BitrateLogProviderMock()
-        mockBitrateLogProvider.events = nil
+        let bitrateDetector = BitrateDetectionService()
         
-        let bitrateDetector = BitrateDetectionService(bitrateLogProvider: mockBitrateLogProvider)
+        // act
+        bitrateDetector.detectBitrateChange()
+        
+        // assert
+        XCTAssertEqual(bitrateDetector.videoBitrate, 0)
+    }
+    
+    func test_detectBitrateChange_should_doNothing_when_accessLogReturnNil() {
+        // arrange
+        let mockAccessLogProvider = AccessLogProviderMock()
+        mockAccessLogProvider.events = nil
+        
+        let bitrateDetector = BitrateDetectionService()
+        bitrateDetector.startMonitoring(accessLogProvider: mockAccessLogProvider)
         
         // act
         bitrateDetector.detectBitrateChange()
@@ -28,10 +40,11 @@ class BitrateDetectionServiceTest: XCTestCase {
     
     func test_detectBitrateChange_should_doNothing_when_bitrateLogIsEmpty() {
         // arrange
-        let mockBitrateLogProvider = BitrateLogProviderMock()
-        mockBitrateLogProvider.events = []
+        let mockAccessLogProvider = AccessLogProviderMock()
+        mockAccessLogProvider.events = []
         
-        let bitrateDetector = BitrateDetectionService(bitrateLogProvider: mockBitrateLogProvider)
+        let bitrateDetector = BitrateDetectionService()
+        bitrateDetector.startMonitoring(accessLogProvider: mockAccessLogProvider)
         
         // act
         bitrateDetector.detectBitrateChange()
@@ -42,13 +55,14 @@ class BitrateDetectionServiceTest: XCTestCase {
     
     func test_detectBitrateChange_should_doNothing_when_bitrateLogDoesNotContainLogWithDurationWatched() {
         // arrange
-        let log = BitrateLogDto()
+        var log = AccessLogDto()
         log.durationWatched = 0
         
-        let mockBitrateLogProvider = BitrateLogProviderMock()
-        mockBitrateLogProvider.events = [log]
+        let mockAccessLogProvider = AccessLogProviderMock()
+        mockAccessLogProvider.events = [log]
         
-        let bitrateDetector = BitrateDetectionService(bitrateLogProvider: mockBitrateLogProvider)
+        let bitrateDetector = BitrateDetectionService()
+        bitrateDetector.startMonitoring(accessLogProvider: mockAccessLogProvider)
         
         // act
         bitrateDetector.detectBitrateChange()
@@ -59,15 +73,17 @@ class BitrateDetectionServiceTest: XCTestCase {
     
     func test_detectBitrateChange_should_doNothing_when_bitrateLogDidNotChange() {
         // arrange
-        let log = BitrateLogDto()
+        var log = AccessLogDto()
         log.durationWatched = 1
         log.indicatedBitrate = 10
         
-        let mockBitrateLogProvider = BitrateLogProviderMock()
-        mockBitrateLogProvider.events = [log]
+        let mockAccessLogProvider = AccessLogProviderMock()
+        mockAccessLogProvider.events = [log]
         
-        let bitrateDetector = BitrateDetectionService(bitrateLogProvider: mockBitrateLogProvider)
-        bitrateDetector.videoBitrate = 10
+        let bitrateDetector = BitrateDetectionService()
+        bitrateDetector.startMonitoring(accessLogProvider: mockAccessLogProvider)
+        bitrateDetector.detectBitrateChange()
+        XCTAssertEqual(bitrateDetector.videoBitrate, 10)
         
         // act
         bitrateDetector.detectBitrateChange()
@@ -78,23 +94,23 @@ class BitrateDetectionServiceTest: XCTestCase {
     
     func test_detectBitrateChange_should_changeBitrateValue() {
         // arrange
-        let log = BitrateLogDto()
+        var log = AccessLogDto()
         log.durationWatched = 1
         log.indicatedBitrate = 20
         
-        let log2 = BitrateLogDto()
+        var log2 = AccessLogDto()
         log2.durationWatched = 2
         log2.indicatedBitrate = 30
         
-        let log3 = BitrateLogDto()
+        var log3 = AccessLogDto()
         log3.durationWatched = 0
         log3.indicatedBitrate = 40
         
-        let mockBitrateLogProvider = BitrateLogProviderMock()
-        mockBitrateLogProvider.events = [log, log2, log3]
+        let mockAccessLogProvider = AccessLogProviderMock()
+        mockAccessLogProvider.events = [log, log2, log3]
         
-        let bitrateDetector = BitrateDetectionService(bitrateLogProvider: mockBitrateLogProvider)
-        bitrateDetector.videoBitrate = 10
+        let bitrateDetector = BitrateDetectionService()
+        bitrateDetector.startMonitoring(accessLogProvider: mockAccessLogProvider)
         
         let videoBitrateChangeExpectation = expectation(description: "should change videoBitrate")
         let kvo = bitrateDetector.observe(\.videoBitrate, options: [.new, .old]) {_,_ in
@@ -112,32 +128,31 @@ class BitrateDetectionServiceTest: XCTestCase {
     
     func test_startMonitoring_should_callDetectBitrateChangeAndChangeValue() {
         // arrange
-        let log = BitrateLogDto()
+        var log = AccessLogDto()
         log.durationWatched = 1
         log.indicatedBitrate = 20
         
-        let log2 = BitrateLogDto()
+        var log2 = AccessLogDto()
         log2.durationWatched = 2
         log2.indicatedBitrate = 30
         
-        let log3 = BitrateLogDto()
+        var log3 = AccessLogDto()
         log3.durationWatched = 0
         log3.indicatedBitrate = 40
         
-        let mockBitrateLogProvider = BitrateLogProviderMock()
-        mockBitrateLogProvider.events = [log]
+        let mockAccessLogProvider = AccessLogProviderMock()
+        mockAccessLogProvider.events = [log]
         
-        let bitrateDetector = BitrateDetectionService(bitrateLogProvider: mockBitrateLogProvider)
-        bitrateDetector.videoBitrate = 10
+        let bitrateDetector = BitrateDetectionService()
         
         var videoBitrateChangeExpectation = expectation(description: "should change videoBitrate first time")
         let kvo = bitrateDetector.observe(\.videoBitrate, options: [.new, .old]) {_,_ in
             videoBitrateChangeExpectation.fulfill()
-            mockBitrateLogProvider.events = [log, log2, log3]
+            mockAccessLogProvider.events = [log, log2, log3]
         }
         
         // act
-        bitrateDetector.startMonitoring()
+        bitrateDetector.startMonitoring(accessLogProvider: mockAccessLogProvider)
         
         // assert
         wait(for: [videoBitrateChangeExpectation], timeout: 1.1)